Add Model Builder hierarchy
