+++ dynawaltz-dsl/src/main/groovy/com/powsybl/dynawaltz/dsl/models/buses/BusGroovyExtension.groovy
@@ -8,10 +8,9 @@
 package com.powsybl.dynawaltz.dsl.models.buses
 
 import com.google.auto.service.AutoService
-import com.powsybl.dsl.DslException
 import com.powsybl.dynamicsimulation.DynamicModel
 import com.powsybl.dynamicsimulation.groovy.DynamicModelGroovyExtension
-import com.powsybl.dynawaltz.dsl.ModelBuilder
+import com.powsybl.dynawaltz.dsl.AbstractDynamicModelBuilder
 import com.powsybl.dynawaltz.dsl.PowsyblDynawoGroovyExtension
 import com.powsybl.dynawaltz.models.buses.StandardBus
 
@@ -32,34 +31,11 @@ class BusGroovyExtension extends PowsyblDynawoGroovyExtension<DynamicModel> impl
         new BusBuilder()
     }
 
-    static class BusBuilder implements ModelBuilder<DynamicModel> {
-        String dynamicModelId
-        String staticId
-        String parameterSetId
-
-        void dynamicModelId(String dynamicModelId) {
-            this.dynamicModelId = dynamicModelId
-        }
-
-        void staticId(String staticId) {
-            this.staticId = staticId
-        }
-
-        void parameterSetId(String parameterSetId) {
-            this.parameterSetId = parameterSetId
-        }
+    static class BusBuilder extends AbstractDynamicModelBuilder {
 
         @Override
         StandardBus build() {
-            if (!staticId) {
-                throw new DslException("'staticId' field is not set")
-            }
-            if (!parameterSetId) {
-                throw new DslException("'parameterSetId' field is not set")
-            }
-            if (!dynamicModelId) {
-                dynamicModelId = staticId
-            }
+            setupBuild()
             new StandardBus(dynamicModelId, staticId, parameterSetId)
         }
     }
