+++ dynawaltz-dsl/src/main/groovy/com/powsybl/dynawaltz/dsl/models/generators/GeneratorFictitiousGroovyExtension.groovy
@@ -7,10 +7,9 @@
 package com.powsybl.dynawaltz.dsl.models.generators
 
 import com.google.auto.service.AutoService
-import com.powsybl.dsl.DslException
 import com.powsybl.dynamicsimulation.DynamicModel
 import com.powsybl.dynamicsimulation.groovy.DynamicModelGroovyExtension
-import com.powsybl.dynawaltz.dsl.ModelBuilder
+import com.powsybl.dynawaltz.dsl.AbstractDynamicModelBuilder
 import com.powsybl.dynawaltz.dsl.PowsyblDynawoGroovyExtension
 import com.powsybl.dynawaltz.models.generators.GeneratorFictitious
 
@@ -29,34 +28,11 @@ class GeneratorFictitiousGroovyExtension extends PowsyblDynawoGroovyExtension<Dy
         new GeneratorFictitiousBuilder()
     }
 
-    static class GeneratorFictitiousBuilder implements ModelBuilder<DynamicModel> {
-        String dynamicModelId
-        String staticId
-        String parameterSetId
-
-        void dynamicModelId(String dynamicModelId) {
-            this.dynamicModelId = dynamicModelId
-        }
-
-        void staticId(String staticId) {
-            this.staticId = staticId
-        }
-
-        void parameterSetId(String parameterSetId) {
-            this.parameterSetId = parameterSetId
-        }
+    static class GeneratorFictitiousBuilder extends AbstractDynamicModelBuilder {
 
         @Override
         GeneratorFictitious build() {
-            if (!staticId) {
-                throw new DslException("'staticId' field is not set")
-            }
-            if (!parameterSetId) {
-                throw new DslException("'parameterSetId' field is not set")
-            }
-            if (!dynamicModelId) {
-                dynamicModelId = staticId
-            }
+            setupBuild()
             new GeneratorFictitious(dynamicModelId, staticId, parameterSetId)
         }
     }
